Moved solr search display helpers for creator, contributor and editor json fields into a module.
