+++ app/services/ubiquity/solr_model_wrapper.rb
@@ -1,0 +1,9 @@
+module Ubiquity
+  class SolrModelWrapper
+    attr_accessor :model, :id
+    def initialize(model, id)
+      @model = model
+      @id = id
+    end
+  end
+end
